hg 3.3.0.20150217
Fix reported issue with Powershell error during install.  Error was:

Cannot convert argument "index", with value: "InstallLocation", for "GetValue" to type "System.Int64": "Cannot convert value "InstallLocation" to type "System.Int64". Error: "Input string was not in a correct format.""

As a result, I'm treating the registry value always as a collection, so it
doesn't break if there are multiple matching entries